refactor: 도메인과 영속성 객체 분리

+++ src/main/java/com/recipe/app/common/utils/JwtService.java
@@ -34,7 +34,7 @@ public class JwtService {
     @Value("${jwt.token-validity-in-ms}")
     private long tokenValidMillisecond;
 
-    public String createJwt(int userId) {
+    public String createJwt(Long userId) {
         Date now = new Date();
         Key key = new SecretKeySpec(Base64.getDecoder().decode(this.secretKey), SignatureAlgorithm.HS256.getJcaName());
         return Jwts.builder()
